Added methods creating order based on cart and changing status order …

+++ src/main/java/com/kodilla/ecommercee/service/OrderService.java
@@ -1,17 +1,25 @@
 package com.kodilla.ecommercee.service;
 
 import com.kodilla.ecommercee.controller.OrderNotFoundException;
+import com.kodilla.ecommercee.domain.Cart;
+import com.kodilla.ecommercee.domain.OrderStatus;
+import com.kodilla.ecommercee.exception.EmptyCartException;
+import com.kodilla.ecommercee.repository.CartRepository;
 import com.kodilla.ecommercee.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import com.kodilla.ecommercee.domain.Order;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 
+@Transactional
 @Service
 @RequiredArgsConstructor
 public class OrderService {
     private final OrderRepository orderRepository;
+    private final CartRepository cartRepository;
 
     public Order saveOrder(final Order order) {
         return orderRepository.save(order);
@@ -34,12 +42,27 @@ public class OrderService {
     }
 
     public void deleteOrder(final Long id) {
-        Order foundOrder = orderRepository.findById(id).orElseThrow(OrderNotFoundException::new);
+        if (!orderRepository.existsById(id)) throw new OrderNotFoundException();
         orderRepository.deleteById(id);
     }
 
+    public void createOrder(final Cart cart) {
+        if (cart.getProducts().isEmpty()) throw new EmptyCartException();
+        Order order = Order.builder()
+                .user(cart.getUser())
+                .products(new ArrayList<>(cart.getProducts()))
+                .status(OrderStatus.NEW).build();
+        orderRepository.save(order);
+        cart.getProducts().clear();
+        cartRepository.save(cart);
 
+    }
 
+    public void changeStatus(final Long orderId, final OrderStatus status) {
+        Order order = orderRepository.findById(orderId).orElseThrow(OrderNotFoundException::new);
+        order.setStatus(status);
+        orderRepository.save(order);
+    }
 }
